Add files via upload
Group Project. Tasks include data analysis, programming, data visualisation and reporting on the findings.

GDP is one of the most important indicator in determining the performance of country economy.The significant factors affecting gdp are population,agriculture,service,industry,health,migration,urban and obesity which are recorded in the dataset.Null hypothesis is that population,agriculture,service,industry,health,migration,urban and obesity are the factors not affecting the gdp per capita of the countries.By analyzing the data we have to find out whether or not these factors affecting gdp per capita of the country.For undergoing the analysis,we have used CIA_Factsheet dataset that records different factors affecting gdp of the countries for the year 2013.In this project, there is a brief description of the factors and validate the stated hypothesis.Preliminary analysis has been conducted based on the Correlation matrix which is showing the relationship of the various variables with the gdp per capita of the country.We can clearly see from the principal component analysis graph that some variables play a very important role in determining the GDP and the GDP Per Capita of a country.